conn_test: implement SUPPORTED in mocked server
Allows customizing what data will be included by TestServer in SUPPORTED
message. This is done by supplying a testSupportedFactory function,
which takes a connection and returns data to be included in SUPPORTED
message for that connection.

Because the driver gets information about the shard assigned to
connection through a SUPPORTED message, this change enables using
the TestServer mock in tests which verify that establishing a connection
per shard works.